% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runShiny.R
\name{runShiny}
\alias{runShiny}
\title{A wrapper function to run Shiny Apps from \code{DroneImageApp}.}
\usage{
runShiny(app = "DroneApp")
}
\arguments{
\item{app}{DroneImgApp}
}
\value{
eviatlas shiny app
}
\description{
Running this function will launch the DroneImgApp Shiny
}
